api/controllers: Rename generate-stream-key to generate-keys
